requirements.txt
litellm>=1.50.0
python-dotenv>=1.0.0
httpx>=0.27.0
